requirements.txt
streamlit
torch
torchvision
numpy
faiss-cpu
Pillow
